requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
alembic==1.12.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0
pytest==7.4.2
httpx==0.24.1
psycopg2-binary==2.9.7
redis==4.6.0
pymongo==4.5.0
celery==5.3.4
sentry-sdk==1.31.0 
loguru==0.6.0
 